Move type definitions and type conversions to different module
